requirements.txt
allure-pytest
dotenv
enum
httpx
pytest 